[internal] Add DP_MSG_INTERNAL_TYPE_SYNC_CANVAS_STATE
This is used to get the state of the canvas at a specific point in the
message stream.

I'm not sure if this particular design is useful upstream, this is
needed here because we need to sync chat with the canvas state. Maybe
there's a more general design for a few of the other callbacks that
could work similarly to this.