Run canvas history tracking in insolation
When `handleCommit` suspends it's possible for another canvas sync
function to run and modify the state of the canvas, for example if
`!reset` is received before the commit is finalized.

This makes all canvas sync functions run in isolation, so the reset
would wait until after the commit is done.