Handle drawpile resets
You need to send an internal message to the paint engine to actually do
a reset.

This allows diffing accross a reset as long as it hasn't reconnected
after.

The drawpile client enters the catchup state after a reset which is not
tracked here, however this does not matter as no chat messages will be
received during that time.